refactor(mesh): set name for steps
