+++ packages/mesh/app/collection/(mutate)/_components/desktop/desktop-edit-collection.tsx
@@ -1,7 +1,7 @@
 'use client'
 
-import DesktopCollectionPicks from './desktop-collection-picks'
-import DesktopMainAction from './desktop-main-action'
+import DesktopCollectionPicks from './collection-picks'
+import DesktopMainAction from './main-action'
 
 export default function DesktopEditCollection() {
   return (
