+++ packages/mesh/app/collection/(mutate)/_components/mobile/mobile-go-next-button.tsx
@@ -1,20 +1,26 @@
 import MobileNavigationButton from '@/components/layout-template/navigation/mobile-navigation/mobile-navigation-button'
 import { useEditCollection } from '@/context/edit-collection'
 
+import { MobielEditCollectionStep } from '../../_types/edit-collection'
+
 export default function MobileGoNextButton() {
-  const { step, setStep, checkMobileStepFullfilled, createCollection } =
-    useEditCollection()
+  const {
+    mobileStepName,
+    setStep,
+    checkMobileStepFullfilled,
+    createCollection,
+  } = useEditCollection()
   const isStepFullfilled = checkMobileStepFullfilled()
 
   const goNextStep = () => {
-    setStep(step + 1)
+    setStep((step) => step + 1)
   }
 
   if (!isStepFullfilled) return null
 
-  switch (step) {
-    case 0:
-    case 1:
+  switch (mobileStepName) {
+    case MobielEditCollectionStep.MobileStep1SelectStories:
+    case MobielEditCollectionStep.MobileStep2SetTitle:
       return (
         <MobileNavigationButton
           text="下一步"
@@ -23,7 +29,7 @@ export default function MobileGoNextButton() {
           color="blue"
         />
       )
-    case 2:
+    case MobielEditCollectionStep.MobileStep3SetSummary:
       return (
         <MobileNavigationButton
           text="建立"
@@ -32,7 +38,9 @@ export default function MobileGoNextButton() {
           color="blue"
         />
       )
-
+    // TODO: implement in phase2
+    // case MobielEditCollectionStep.MobileStep4SortStories:
+    //   return null
     default:
       return null
   }
